pyproject.toml
[tool.poetry]
name = "books-api"
version = "0.1.0"
description = ""
authors = ["Ivan Zugnoni"]
readme = "README.md"
packages = [{include = "books_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-model-utils = "^4.3.1"
drf-nested-routers = "^0.93.4"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
black = "^23.1.0"
pylint = "^2.17.0"
ipython = "^8.11.0"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
